refactor: UserRepository를 interface로 수정
- 구현체 클래스  MemoryUserRepository로 이름 변경
