+++ src/main/java/MvcConfig.java
@@ -1,6 +1,5 @@
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
 public class MvcConfig implements WebMvcConfigurer {
