+++ src/main/java/com/kakao/cafe/domain/users/UserRepository.java
@@ -1,0 +1,10 @@
+package com.kakao.cafe.domain.users;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface UserRepository {
+	public List<Users> findAll();
+	public void save(Users user);
+	public Optional<Users> findByUserId = null;
+}
