+++ src/main/java/com/kakao/cafe/domain/users/Users.java
@@ -1,4 +1,4 @@
-package com.kakao.cafe.domain.user;
+package com.kakao.cafe.domain.users;
 
 public class Users {
