+++ src/main/java/com/kakao/cafe/service/users/UserService.java
@@ -1,14 +1,14 @@
 package com.kakao.cafe.service.users;
 
-import com.kakao.cafe.domain.user.Users;
-import com.kakao.cafe.domain.user.UserRepository;
+import com.kakao.cafe.domain.users.Users;
+import com.kakao.cafe.domain.users.MemoryUserRepository;
 import java.util.List;
 import org.springframework.stereotype.Service;
 
 @Service
 public class UserService {
 
-	UserRepository userRepository = new UserRepository();
+	MemoryUserRepository userRepository = new MemoryUserRepository();
 
 	public void save(Users user) {
 		userRepository.save(user);
